Move `GovernanceActionProtocolParametersUpdateCmdArgs era` to `Cardano.CLI.Compatible.Governance.Types`
