+++ cardano-cli/src/Cardano/CLI/EraBased/Governance/Actions/Command.hs
@@ -17,7 +17,6 @@ module Cardano.CLI.EraBased.Governance.Actions.Command
   , UpdateProtocolParametersConwayOnwards (..)
   , UpdateProtocolParametersPreConway (..)
   , GovernanceActionHardforkInitCmdArgs (..)
-  , CostModelsFile (..)
   , renderGovernanceActionCmds
   )
 where
@@ -26,6 +25,7 @@ import Cardano.Api
 import Cardano.Api.Experimental qualified as Exp
 import Cardano.Api.Ledger qualified as L
 
+import Cardano.CLI.Compatible.Governance.Types
 import Cardano.CLI.Type.Common
 import Cardano.CLI.Type.Key
 
@@ -106,24 +106,6 @@ data GovernanceActionCreateNoConfidenceCmdArgs era
   }
   deriving Show
 
-data GovernanceActionProtocolParametersUpdateCmdArgs era
-  = GovernanceActionProtocolParametersUpdateCmdArgs
-  { uppShelleyBasedEra :: !(ShelleyBasedEra era)
-  , uppPreConway :: !(Maybe (UpdateProtocolParametersPreConway era))
-  , uppConwayOnwards :: !(Maybe (UpdateProtocolParametersConwayOnwards era))
-  , uppNewPParams :: !(EraBasedProtocolParametersUpdate era)
-  -- ^ New parameters to be proposed. From Alonzo onwards, the type
-  -- 'EraBasedProtocolParametersUpdate' also contains cost models. Since all
-  -- other protocol parameters are read from command line arguments, whereas
-  -- the cost models are read from a file, we separate the cost models from
-  -- the rest of the protocol parameters to ease parsing.
-  , uppCostModelsFile :: !(Maybe (CostModelsFile era))
-  -- ^ The new cost models proposed. See the comment at 'uppNewPParams' for
-  -- why this is a separate field.
-  , uppFilePath :: !(File () Out)
-  }
-  deriving Show
-
 data GovernanceActionTreasuryWithdrawalCmdArgs era
   = GovernanceActionTreasuryWithdrawalCmdArgs
   { era :: !(Exp.Era era)
@@ -162,37 +144,6 @@ data GovernanceActionViewCmdArgs era
   , mOutFile :: !(Maybe (File () Out))
   }
   deriving Show
-
-data UpdateProtocolParametersConwayOnwards era
-  = UpdateProtocolParametersConwayOnwards
-  { era :: !(Exp.Era era)
-  , networkId :: !L.Network
-  , deposit :: !Lovelace
-  , returnAddr :: !StakeIdentifier
-  , proposalUrl :: !ProposalUrl
-  , proposalHash :: !(L.SafeHash L.AnchorData)
-  , checkProposalHash :: !(MustCheckHash ProposalUrl)
-  , governanceActionId :: !(Maybe L.GovActionId)
-  , constitutionScriptHash :: !(Maybe ScriptHash)
-  }
-
-data CostModelsFile era
-  = CostModelsFile
-  { eon :: !(AlonzoEraOnwards era)
-  , costModelsFile :: !(File L.CostModels In)
-  }
-  deriving Show
-
-deriving instance Show (UpdateProtocolParametersConwayOnwards era)
-
-data UpdateProtocolParametersPreConway era
-  = UpdateProtocolParametersPreConway
-  { eon :: !(ShelleyToBabbageEra era)
-  , expiryEpoch :: !EpochNo
-  , genesisVerificationKeys :: ![VerificationKeyFile In]
-  }
-
-deriving instance Show (UpdateProtocolParametersPreConway era)
 
 renderGovernanceActionCmds :: GovernanceActionCmds era -> Text
 renderGovernanceActionCmds =
